CompatiblePosts tests are ok, and all backend test are ok for now too

+++ backend/src/main/java/com/bloodify/backend/UserRequests/service/interfaces/CompatiblePosts.java
@@ -6,5 +6,5 @@ import com.bloodify.backend.UserRequests.model.entities.Post;
 import java.util.List;
 
 public interface CompatiblePosts {
-    public List<Post> allPostsMatching (User user, double threshold);
+    public List<Post> allPostsMatching (User user, int threshold);
 }
